requirements.txt
flask==2.3.2
flask_cors==4.0.0
flask_bootstrap4==4.0.2